boards/requirements.txt
asgiref==3.4.1
biopython==1.79
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.6
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
idna==3.2
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mock==1.0.1
numpy==1.19.5
oauthlib==3.1.1
Pillow==8.3.1
pkg_resources==0.0.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pysam==0.16.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
RUST==0.1.1
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
